Bump upper version limit for the stdlib dependency (#17)
In addition rework fixture file to use forge release when runnign
rspec-puppet tests. This automatically fixes the old repo URL for the
EPEL module, which still pointed to the old stahnma repo. No error
occured previsouly as this URL was redirect to the version in org
voxpupuli, where it officially lives now.